separate wasm-opt out from wasm-shell: opt optimizes, shell runs wast shell tests
