+++ CMakeLists.txt
@@ -120,6 +120,17 @@ SET_PROPERTY(TARGET wasm-shell PROPERTY CXX_STANDARD 11)
 SET_PROPERTY(TARGET wasm-shell PROPERTY CXX_STANDARD_REQUIRED ON)
 INSTALL(TARGETS wasm-shell DESTINATION bin)
 
+SET(wasm-opt_SOURCES
+  src/tools/wasm-opt.cpp
+  src/wasm.cpp
+)
+ADD_EXECUTABLE(wasm-opt
+               ${wasm-opt_SOURCES})
+TARGET_LINK_LIBRARIES(wasm-opt asmjs emscripten-optimizer ${all_passes} support)
+SET_PROPERTY(TARGET wasm-opt PROPERTY CXX_STANDARD 11)
+SET_PROPERTY(TARGET wasm-opt PROPERTY CXX_STANDARD_REQUIRED ON)
+INSTALL(TARGETS wasm-opt DESTINATION bin)
+
 SET(asm2wasm_SOURCES
   src/tools/asm2wasm.cpp
   src/wasm.cpp
